# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Server API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/admin/users": {
            "get": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "List all users with optional pagination parameters and search functionality",
                "tags": [
                    "admin"
                ],
                "summary": "List users with pagination and optional search",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_controllers.AdminListUsersResponse"
                        }
                    },
                    "400": {
                        "description": "error: Invalid page or page size",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "error: failed to list users",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/users/{id}": {
            "delete": {
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "csrf": []
                    }
                ],
                "description": "Remove a user from the database by providing their ID",
                "tags": [
                    "admin"
                ],
                "summary": "Remove a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: user removed",
                        "schema": {
                            "$ref": "#/definitions/api_controllers.AdminRemoveUserResponse"
                        }
                    },
                    "500": {
                        "description": "error: failed to remove user",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/users/{id}/role": {
            "put": {
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "csrf": []
                    }
                ],
                "description": "Update the role of a user identified by their ID with the provided role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update user role by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Role",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_controllers.AdminUpdateUserRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: user role updated, user: updated user details",
                        "schema": {
                            "$ref": "#/definitions/api_controllers.AdminUpdateUserRoleResponse"
                        }
                    },
                    "400": {
                        "description": "error: invalid input or invalid role",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "error: user not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "error: failed to update user role",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_controllers.AdminListUsersResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "api_controllers.AdminRemoveUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api_controllers.AdminUpdateUserRoleRequest": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                }
            }
        },
        "api_controllers.AdminUpdateUserRoleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Role": {
            "type": "string",
            "enum": [
                "user",
                "admin",
                "disabled"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleAdmin",
                "RoleDisabled"
            ]
        },
        "models.User": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "totp_enabled": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "csrf": {
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}